Clean up some sub-optimal slice handling
Clippy noticed that update_from looked like a manual copy_from_slice,
which alerted me to sum_into and mul_into also being suboptimal. With
the new changes, the generated code is much cleaner: it lacks an extra
"unnecessary" branch and is about 30% shorter (for those methods).

I also think the code is a bit easier to read :)